backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
openai-whisper==20231117
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
librosa==0.10.1
scipy==1.11.4
